style the sign up page
